Fixed some changes that were made to Joystick's definition so that linux matches them.

+++ source/linux/Joystick.cpp
@@ -113,7 +113,7 @@ namespace XPG
 
         _axisStates = new Int32[_numAxes];
         _buttonStates = new bool[_numButtons];
-        _hatStates = new Int32[_numHats];
+        _hatStates = new Int16[_numHats];
     }
 
 
@@ -126,7 +126,7 @@ namespace XPG
 
         _axisStates = new Int32[_numAxes];
         _buttonStates = new bool[_numButtons];
-        _hatStates = new Int32[_numHats];
+        _hatStates = new Int16[_numHats];
 
         _axisMinimums = 0;
         _axisMaximums = 0;
@@ -246,7 +246,7 @@ namespace XPG
 
         _axisStates = new Int32[_numAxes];
         _buttonStates = new bool[_numButtons];
-        _hatStates = new Int32[_numHats];
+        _hatStates = new Int16[_numHats];
 
         for (UInt32 i = 0; i < _numAxes; ++i)
         {
